# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "E-Commerce Demo API",
    "version": "1.0.0",
    "description": "花卉電商網站 REST API"
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "sessionAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Session-Id"
      }
    },
...
  },
  "paths": {
...
    "/api/orders": {
      "post": {
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "recipientName": {
                    "type": "string",
                    "minLength": 1
                  },
                  "recipientEmail": {
                    "type": "string",
                    "minLength": 1,
                    "format": "email"
                  },
                  "recipientAddress": {
                    "type": "string",
                    "minLength": 1
                  },
                  "shippingMethod": {
                    "type": "string",
                    "enum": [
                      "home_delivery",
                      "convenience_store"
                    ]
                  },
                  "isRemoteArea": {
                    "type": "boolean",
                    "default": false
                  },
                  "isRushDelivery": {
                    "type": "boolean",
                    "default": false
                  }
                },
                "required": [
                  "recipientName",
                  "recipientEmail",
                  "recipientAddress",
                  "shippingMethod"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "訂單建立成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "order_no": {
                          "type": "string"
                        },
                        "subtotal": {
                          "type": "number",
                          "nullable": true
                        },
                        "shipping_method": {
                          "type": "string",
                          "nullable": true
                        },
                        "shipping_fee": {
                          "type": "number",
                          "nullable": true
                        },
                        "is_remote_area": {
                          "anyOf": [
                            {
                              "type": "number",
                              "enum": [
                                0
                              ]
                            },
                            {
                              "type": "number",
                              "enum": [
                                1
                              ]
                            },
                            {
                              "nullable": true
                            }
                          ]
                        },
                        "is_rush_delivery": {
                          "anyOf": [
                            {
                              "type": "number",
                              "enum": [
                                0
                              ]
                            },
                            {
                              "type": "number",
                              "enum": [
                                1
                              ]
                            },
                            {
                              "nullable": true
                            }
                          ]
                        },
                        "total_amount": {
                          "type": "number"
                        },
                        "status": {
                          "type": "string"
                        },
                        "created_at": {
                          "type": "string"
                        },
                        "items": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "product_name": {
                                "type": "string"
                              },
                              "product_price": {
                                "type": "number"
                              },
                              "quantity": {
                                "type": "number"
                              }
                            },
                            "required": [
                              "product_name",
                              "product_price",
                              "quantity"
                            ]
                          }
                        }
                      },
                      "required": [
                        "id",
                        "order_no",
                        "subtotal",
                        "shipping_method",
                        "shipping_fee",
                        "is_remote_area",
                        "is_rush_delivery",
                        "total_amount",
                        "status",
                        "created_at",
                        "items"
                      ]
                    },
                    "error": {
                      "type": "string",
                      "nullable": true
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "data",
                    "error",
                    "message"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "購物車為空或庫存不足或收件資訊缺失或配送方式無效"
          },
          "401": {
            "description": "未登入或 token 無效"
          }
        }
      },
      "get": {
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "orders": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "order_no": {
                                "type": "string"
                              },
                              "total_amount": {
                                "type": "number"
                              },
                              "status": {
                                "type": "string"
                              },
                              "created_at": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "id",
                              "order_no",
                              "total_amount",
                              "status",
                              "created_at"
                            ]
                          }
                        }
                      },
                      "required": [
                        "orders"
                      ]
                    },
                    "error": {
                      "type": "string",
                      "nullable": true
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "data",
                    "error",
                    "message"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "未登入或 token 無效"
          }
        }
      }
    },
    "/api/orders/{id}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "order_no": {
                          "type": "string"
                        },
                        "user_id": {
                          "type": "string"
                        },
                        "recipient_name": {
                          "type": "string"
                        },
                        "recipient_email": {
                          "type": "string"
                        },
                        "recipient_address": {
                          "type": "string"
                        },
                        "subtotal": {
                          "type": "number",
                          "nullable": true
                        },
                        "shipping_method": {
                          "type": "string",
                          "nullable": true
                        },
                        "shipping_fee": {
                          "type": "number",
                          "nullable": true
                        },
                        "is_remote_area": {
                          "anyOf": [
                            {
                              "type": "number",
                              "enum": [
                                0
                              ]
                            },
                            {
                              "type": "number",
                              "enum": [
                                1
                              ]
                            },
                            {
                              "nullable": true
                            }
                          ]
                        },
                        "is_rush_delivery": {
                          "anyOf": [
                            {
                              "type": "number",
                              "enum": [
                                0
                              ]
                            },
                            {
                              "type": "number",
                              "enum": [
                                1
                              ]
                            },
                            {
                              "nullable": true
                            }
                          ]
                        },
                        "total_amount": {
                          "type": "number"
                        },
                        "status": {
                          "type": "string"
                        },
                        "merchant_trade_no": {
                          "type": "string",
                          "nullable": true
                        },
                        "created_at": {
                          "type": "string"
                        },
                        "items": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "order_id": {
                                "type": "string"
                              },
                              "product_id": {
                                "type": "string"
                              },
                              "product_name": {
                                "type": "string"
                              },
                              "product_price": {
                                "type": "number"
                              },
                              "quantity": {
                                "type": "number"
                              }
                            },
                            "required": [
                              "id",
                              "order_id",
                              "product_id",
                              "product_name",
                              "product_price",
                              "quantity"
                            ]
                          }
                        }
                      },
                      "required": [
                        "id",
                        "order_no",
                        "user_id",
                        "recipient_name",
                        "recipient_email",
                        "recipient_address",
                        "subtotal",
                        "shipping_method",
                        "shipping_fee",
                        "is_remote_area",
                        "is_rush_delivery",
                        "total_amount",
                        "status",
                        "merchant_trade_no",
                        "created_at",
                        "items"
                      ]
                    },
                    "error": {
                      "type": "string",
                      "nullable": true
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "data",
                    "error",
                    "message"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "未登入或 token 無效"
          },
          "404": {
            "description": "訂單不存在"
          }
        }
      }
    },
    "/api/orders/{id}/pay": {
      "patch": {
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "action": {
                    "type": "string",
                    "enum": [
                      "success",
                      "fail"
                    ]
                  }
                },
                "required": [
                  "action"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "付款狀態更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "order_no": {
                          "type": "string"
                        },
                        "subtotal": {
                          "type": "number",
                          "nullable": true
                        },
                        "shipping_method": {
                          "type": "string",
                          "nullable": true
                        },
                        "shipping_fee": {
                          "type": "number",
                          "nullable": true
                        },
                        "is_remote_area": {
                          "anyOf": [
                            {
                              "type": "number",
                              "enum": [
                                0
                              ]
                            },
                            {
                              "type": "number",
                              "enum": [
                                1
                              ]
                            },
                            {
                              "nullable": true
                            }
                          ]
                        },
                        "is_rush_delivery": {
                          "anyOf": [
                            {
                              "type": "number",
                              "enum": [
                                0
                              ]
                            },
                            {
                              "type": "number",
                              "enum": [
                                1
                              ]
                            },
                            {
                              "nullable": true
                            }
                          ]
                        },
                        "total_amount": {
                          "type": "number"
                        },
                        "status": {
                          "type": "string"
                        },
                        "created_at": {
                          "type": "string"
                        },
                        "items": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "product_name": {
                                "type": "string"
                              },
                              "product_price": {
                                "type": "number"
                              },
                              "quantity": {
                                "type": "number"
                              }
                            },
                            "required": [
                              "product_name",
                              "product_price",
                              "quantity"
                            ]
                          }
                        }
                      },
                      "required": [
                        "id",
                        "order_no",
                        "subtotal",
                        "shipping_method",
                        "shipping_fee",
                        "is_remote_area",
                        "is_rush_delivery",
                        "total_amount",
                        "status",
                        "created_at",
                        "items"
                      ]
                    },
                    "error": {
                      "type": "string",
                      "nullable": true
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "data",
                    "error",
                    "message"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "action 無效或訂單狀態不是 pending"
          },
          "401": {
            "description": "未登入或 token 無效"
          },
          "404": {
            "description": "訂單不存在"
          }
        }
      }
    },
    "/api/orders/{id}/check-payment": {
      "post": {
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "查詢成功"
          },
          "400": {
            "description": "訂單狀態不是 pending"
          },
          "401": {
            "description": "未登入或 token 無效"
          },
          "404": {
            "description": "訂單不存在"
          }
        }
      }
    },
...
  }
}